Updated api gateway to handle CORS and have a user authentication route also as a middleware
